refactor(station-manager): update uow classes hierarchy
