+++ backend/station_manager/app/src/main/java/com/StationManager/app/services/unitofwork/IUnitOfWork.java
@@ -1,6 +1,7 @@
 package com.StationManager.app.services.unitofwork;
 
 import com.StationManager.app.domain.events.Event;
+import com.StationManager.app.storage.repository.*;
 
 import java.util.List;
 
@@ -8,4 +9,10 @@ public interface IUnitOfWork extends AutoCloseable {
     List<Event> collectNewEvents();
     void commit();
     void rollback();
+
+    IClientRepository getClientRepository();
+    IHallRepository getHallRepository();
+    IPrivilegyRepository getPrivilegyRepository();
+    ITicketOfficeRepository getTicketOfficeRepository();
+    ITrainStationRepository getTrainStationRepository();
 }
