+++ backend/station_manager/app/src/main/java/com/StationManager/app/services/unitofwork/PostgresUnitOfWork.java
@@ -11,40 +11,18 @@ import java.util.List;
 
 public class PostgresUnitOfWork extends UnitOfWork {
     public Session session;
-//
-    public final IClientRepository clientRepository;
-    public final IHallRepository hallRepository;
-    public final IPrivilegyRepository privilegyRepository;
-    public final ITicketOfficeRepository ticketOfficeRepository;
-    public final ITrainStationRepository trainStationRepository;
 
     public PostgresUnitOfWork() {
         // this.session = PgSessionFactory.openSession();  // TODO: replace with real session factory
         //noinspection resource - session is closed in close() method
         this.session = new Configuration().configure().buildSessionFactory().openSession();
 
-        // replace wiht correct repos
+        // replace with correct repos
         this.clientRepository = new InMemoryClientRepository();
         this.hallRepository = new InMemoryHallRepository();
         this.privilegyRepository = new InMemoryPrivilegyRepository();
         this.ticketOfficeRepository = new InMemoryTicketOfficeRepository();
         this.trainStationRepository = new InMemoryTrainStationRepository();
-    }
-
-    @Override
-    public List<Event> collectNewEvents() {
-        var events = new ArrayList<Event>();
-        for (var hall: hallRepository.getSeen()) {
-            while (!hall.events.isEmpty()) {
-                events.add(hall.events.poll());
-            }
-        }
-        for (var ticketOffice: ticketOfficeRepository.getSeen()) {
-            while (!ticketOffice.events.isEmpty()) {
-                events.add(ticketOffice.events.poll());
-            }
-        }
-        return events.stream().sorted().toList();
     }
 
     @Override
